refactor(qiskit): refactor of Qiskit adapter structure and tests
