+++ packages/devqubit-qiskit/src/devqubit_qiskit/results.py
@@ -17,6 +17,11 @@ from devqubit_engine.uec.models.result import ResultType
 from devqubit_engine.utils.serialization import to_jsonable
 
 
+# =============================================================================
+# Result Type Detection
+# =============================================================================
+
+
 def detect_result_type(result: Any) -> ResultType:
     """
     Detect the result type from a Qiskit result object.
@@ -46,15 +51,6 @@ def detect_result_type(result: Any) -> ResultType:
     except Exception:
         pass
 
-    # Check for unitary/density matrix (simulator)
-    try:
-        if hasattr(result, "get_unitary") and callable(result.get_unitary):
-            unitary = result.get_unitary()
-            if unitary is not None:
-                return ResultType.DENSITY_MATRIX
-    except Exception:
-        pass
-
     # Check for quasi-distributions (Runtime Sampler)
     try:
         if extract_quasi_distributions(result) is not None:
@@ -78,6 +74,11 @@ def detect_result_type(result: Any) -> ResultType:
         pass
 
     return ResultType.OTHER
+
+
+# =============================================================================
+# Counts Extraction
+# =============================================================================
 
 
 def normalize_result_counts(result: Any) -> dict[str, Any]:
@@ -111,91 +112,74 @@ def normalize_result_counts(result: Any) -> dict[str, Any]:
     if result is None:
         return output
 
-    # Best-effort number of experiments
-    num_experiments: int | None = None
-    try:
-        if hasattr(result, "results") and result.results is not None:
-            num_experiments = len(result.results)
-    except Exception:
-        pass
+    # Determine number of experiments
+    num_experiments = _get_num_experiments(result)
 
-    # If we cannot determine experiments, try a single get_counts()
+    # Single experiment fallback
     if not num_experiments:
-        try:
-            if hasattr(result, "get_counts") and callable(result.get_counts):
-                counts = result.get_counts()
-                counts_dict = to_jsonable(counts)
-                shots = (
-                    sum(counts_dict.values()) if isinstance(counts_dict, dict) else None
-                )
-                output["experiments"].append(
-                    {
-                        "index": 0,
-                        "counts": counts_dict,
-                        "shots": shots,
-                    }
-                )
-        except Exception:
-            pass
+        exp = _extract_single_experiment(result)
+        if exp:
+            output["experiments"].append(exp)
         return output
 
     # Multiple experiments
     for idx in range(int(num_experiments)):
-        try:
-            counts = result.get_counts(idx)
-            counts_dict = to_jsonable(counts)
-            shots = sum(counts_dict.values()) if isinstance(counts_dict, dict) else None
-
-            # Try to get experiment name
-            name = None
-            try:
-                if hasattr(result.results[idx], "header"):
-                    name = getattr(result.results[idx].header, "name", None)
-            except Exception:
-                pass
-
-            exp_data: dict[str, Any] = {
-                "index": idx,
-                "counts": counts_dict,
-                "shots": shots,
-            }
-            if name:
-                exp_data["name"] = str(name)
-
-            output["experiments"].append(exp_data)
-        except Exception:
-            continue
+        exp = _extract_experiment_by_index(result, idx)
+        if exp:
+            output["experiments"].append(exp)
 
     return output
 
 
-def _safe_getattr(obj: Any, attr: str, converter: type | None = None) -> Any | None:
-    """
-    Safely get and optionally convert an attribute value.
-
-    Parameters
-    ----------
-    obj : Any
-        Object to get attribute from.
-    attr : str
-        Attribute name.
-    converter : type, optional
-        Type to convert value to (str, bool, float).
-
-    Returns
-    -------
-    Any or None
-        Converted value or None if attribute missing or conversion fails.
-    """
+def _get_num_experiments(result: Any) -> int | None:
+    """Get number of experiments from result."""
     try:
-        if not hasattr(obj, attr):
-            return None
-        val = getattr(obj, attr)
-        if converter is not None:
-            return converter(val)
-        return val
+        if hasattr(result, "results") and result.results is not None:
+            return len(result.results)
+    except Exception:
+        pass
+    return None
+
+
+def _extract_single_experiment(result: Any) -> dict[str, Any] | None:
+    """Extract counts from a single-experiment result."""
+    try:
+        if hasattr(result, "get_counts") and callable(result.get_counts):
+            counts = result.get_counts()
+            counts_dict = to_jsonable(counts)
+            shots = sum(counts_dict.values()) if isinstance(counts_dict, dict) else None
+            return {"index": 0, "counts": counts_dict, "shots": shots}
+    except Exception:
+        pass
+    return None
+
+
+def _extract_experiment_by_index(result: Any, idx: int) -> dict[str, Any] | None:
+    """Extract counts for a specific experiment index."""
+    try:
+        counts = result.get_counts(idx)
+        counts_dict = to_jsonable(counts)
+        shots = sum(counts_dict.values()) if isinstance(counts_dict, dict) else None
+
+        exp_data: dict[str, Any] = {"index": idx, "counts": counts_dict, "shots": shots}
+
+        # Try to get experiment name
+        try:
+            if hasattr(result.results[idx], "header"):
+                name = getattr(result.results[idx].header, "name", None)
+                if name:
+                    exp_data["name"] = str(name)
+        except Exception:
+            pass
+
+        return exp_data
     except Exception:
         return None
+
+
+# =============================================================================
+# Result Metadata
+# =============================================================================
 
 
 def extract_result_metadata(result: Any) -> dict[str, Any]:
@@ -218,22 +202,39 @@ def extract_result_metadata(result: Any) -> dict[str, Any]:
     if result is None:
         return metadata
 
-    # Define attributes to extract: (attr_name, output_key, converter)
     attrs = [
-        ("backend_name", "backend_name", str),
-        ("job_id", "job_id", str),
-        ("success", "success", bool),
-        ("status", "status", str),
-        ("date", "date", str),
-        ("time_taken", "time_taken", float),
+        ("backend_name", str),
+        ("job_id", str),
+        ("success", bool),
+        ("status", str),
+        ("date", str),
+        ("time_taken", float),
     ]
 
-    for attr, key, converter in attrs:
+    for attr, converter in attrs:
         val = _safe_getattr(result, attr, converter)
         if val is not None:
-            metadata[key] = val
+            metadata[attr] = val
 
     return metadata
+
+
+def _safe_getattr(obj: Any, attr: str, converter: type | None = None) -> Any | None:
+    """Safely get and optionally convert an attribute value."""
+    try:
+        if not hasattr(obj, attr):
+            return None
+        val = getattr(obj, attr)
+        if converter is not None:
+            return converter(val)
+        return val
+    except Exception:
+        return None
+
+
+# =============================================================================
+# Quasi-Distribution Extraction
+# =============================================================================
 
 
 def extract_quasi_distributions(result: Any) -> list[dict[str, float]] | None:
@@ -264,7 +265,6 @@ def extract_quasi_distributions(result: Any) -> list[dict[str, float]] | None:
 
     quasi_dists: list[dict[str, float]] = []
 
-    # Try direct quasi_dists attribute (Runtime Sampler)
     try:
         if hasattr(result, "quasi_dists") and result.quasi_dists is not None:
             for qd in result.quasi_dists:
@@ -273,7 +273,6 @@ def extract_quasi_distributions(result: Any) -> list[dict[str, float]] | None:
                 ):
                     quasi_dists.append(dict(qd.binary_probabilities()))
                 elif isinstance(qd, dict):
-                    # Convert integer keys to binary strings
                     if qd:
                         num_bits = max(len(bin(k)) - 2 for k in qd.keys())
                     else:
@@ -328,142 +327,9 @@ def extract_expectation_values(
     return None
 
 
-def extract_statevector(result: Any) -> dict[str, Any] | None:
-    """
-    Extract statevector from a Qiskit simulator result.
-
-    Parameters
-    ----------
-    result : Any
-        Qiskit result object (typically from Aer simulator).
-
-    Returns
-    -------
-    dict or None
-        Statevector data with amplitude information, or None if
-        not available.
-
-    Notes
-    -----
-    Statevectors are only available from simulator backends and
-    can be very large for many qubits. The returned dict contains:
-
-    - "num_qubits": Number of qubits
-    - "dim": Statevector dimension (2^n)
-    - "data_type": "complex128" or similar
-    - "amplitudes": List of [real, imag] pairs (truncated for large states)
-    """
-    try:
-        if hasattr(result, "get_statevector"):
-            sv = result.get_statevector()
-            if sv is not None:
-                import numpy as np
-
-                data = np.asarray(sv.data)
-                num_qubits = int(np.log2(len(data)))
-
-                # For large statevectors, only store metadata
-                if num_qubits > 10:
-                    return {
-                        "num_qubits": num_qubits,
-                        "dim": len(data),
-                        "data_type": str(data.dtype),
-                        "truncated": True,
-                        "note": "Statevector too large for inline storage",
-                    }
-
-                # Store full statevector for small states
-                amplitudes = [[float(x.real), float(x.imag)] for x in data]
-                return {
-                    "num_qubits": num_qubits,
-                    "dim": len(data),
-                    "data_type": str(data.dtype),
-                    "amplitudes": amplitudes,
-                }
-    except Exception:
-        pass
-
-    return None
-
-
-def extract_unitary(result: Any) -> dict[str, Any] | None:
-    """
-    Extract unitary matrix from a Qiskit simulator result.
-
-    Parameters
-    ----------
-    result : Any
-        Qiskit result object (typically from Aer UnitarySimulator).
-
-    Returns
-    -------
-    dict or None
-        Unitary matrix metadata, or None if not available.
-
-    Notes
-    -----
-    Full unitary matrices are typically too large to store inline.
-    This function returns metadata about the unitary.
-    """
-    try:
-        if hasattr(result, "get_unitary"):
-            unitary = result.get_unitary()
-            if unitary is not None:
-                import numpy as np
-
-                data = np.asarray(unitary.data)
-                dim = data.shape[0]
-                num_qubits = int(np.log2(dim))
-
-                return {
-                    "num_qubits": num_qubits,
-                    "dim": dim,
-                    "data_type": str(data.dtype),
-                    "shape": list(data.shape),
-                    "note": "Full unitary matrix stored separately as artifact",
-                }
-    except Exception:
-        pass
-
-    return None
-
-
-def extract_density_matrix(result: Any) -> dict[str, Any] | None:
-    """
-    Extract density matrix from a Qiskit simulator result.
-
-    Parameters
-    ----------
-    result : Any
-        Qiskit result object (typically from Aer DensityMatrixSimulator).
-
-    Returns
-    -------
-    dict or None
-        Density matrix metadata, or None if not available.
-    """
-    try:
-        if hasattr(result, "data"):
-            data_dict = result.data()
-            if isinstance(data_dict, dict) and "density_matrix" in data_dict:
-                import numpy as np
-
-                dm = data_dict["density_matrix"]
-                dm_data = np.asarray(dm)
-                dim = dm_data.shape[0]
-                num_qubits = int(np.log2(dim))
-
-                return {
-                    "num_qubits": num_qubits,
-                    "dim": dim,
-                    "data_type": str(dm_data.dtype),
-                    "shape": list(dm_data.shape),
-                    "purity": float(np.real(np.trace(dm_data @ dm_data))),
-                }
-    except Exception:
-        pass
-
-    return None
+# =============================================================================
+# Primitive Result Normalization
+# =============================================================================
 
 
 def normalize_primitive_result(result: Any) -> dict[str, Any]:
